Pathfinding sorta works now
There's a case where the player is on one side of the map, I click on
the other side, and the path somehow = 1. Then of course the player
proceeds to beeline.

Check to make sure the neighbors are being set properly. If that's not
it, then, debug through the algo I guess.